Bash files: fix backwards cascade; tidier formatting; more pedantry
I went a little crazy with the boolean sequencing syntax and trying to
turn way too many things into a loop. The path setup doesn't really need
that. I'm fine with a series of one-shot calls if it means we don't have
to stand up an array that we destroy 3 milliseconds later.

Same with the variable exports. That syntax is actually harder to read
once you key more than one export off of a boolean check.

Finally, got the bash startup files set such that .profile loads on
login, and .bashrc loads in every interactive shell. I had previously
configured the files such that .bashrc was somehow always loaded, and
.profile was only loaded when called by .bashrc. Interestingly, never
saw any weird side effects from this configuration as far as I could
tell, but for whatever reason, it provides a bit of creature comfort to
have confirmed that my system is capable of performing normally without
non-standard configuration acrobatics.

Also moved the functions and aliases closer together in .bashrc; fits
better with standard script architecture to move all the declarations to
the top, even if they're not directly used in this file.