.bashrc: remove unused exported functions
Exporting functions is not a good way to configure a durable environment
ime. Instead, we should just be adding functions directly to scripts or
sourcing them from files if they're ubiquitous.

The only function of this kind that was actually in use in my
environment, is `is_running`, which checks for the respective command
before attempting to run it.

This commit arguably only solves half the problem with the bspwm config,
as we need more graceful fallback handling and meaningful message output
in the event that crucial programs are missing. What I probably should
do from here is think about a PKGBUILD to construct a (for my purposes)
"complete" bspwm env.