ssh-pass.sh: remove pw dependency
I've recently refactored several scripts to bash functions, including
`pw`. This is intentional, because I don't want simple wrapper/alias
commands to become dependencies in my workflow elsewhere - they're just
intended to be convenience functions from the standpoint of keystroke
economy.

This is one of the remnants of such a dependency that missed until it
borked my login flow. For a use case like ssh-pass.sh, easier is to
wrap the command with its assumptions in a script-local function.