bspwm: update geometry
Trying an interim solution towards a true BSP-driven approach with two
main windows in the main node and a stack node that pushes new windows
to the top.

That's going to take a while to get right, so for now we fake it by
letting the browser be a floating window and padding the tiled space
accordingly.