tym: bring back transparent bg, handle with a function
The just-world implementation of this would use a module with a function
that could just be pulled in , but I've tried every permutation of
`require` calls with no success. Not sure if it's a skill issue or just
comes down to the fact that application-dependent Lua shouldn't be
expected to handle external modules.

A template may be a better way to go here.